+++ app/src/main/java/com/nick_sib/popularlibraries/mvp/model/repo/retrofit/RetrofitTheUserRepos.kt
@@ -1,10 +1,9 @@
 package com.nick_sib.popularlibraries.mvp.model.repo.retrofit
 
 import com.nick_sib.popularlibraries.mvp.model.api.IDataSource
+import com.nick_sib.popularlibraries.mvp.model.cache.IGithubReposCache
 import com.nick_sib.popularlibraries.mvp.model.entity.GithubUser
 import com.nick_sib.popularlibraries.mvp.model.entity.GithubRepository
-import com.nick_sib.popularlibraries.mvp.model.entity.room.Database
-import com.nick_sib.popularlibraries.mvp.model.entity.room.RoomGithubRepository
 import com.nick_sib.popularlibraries.mvp.model.network.INetworkStatus
 import com.nick_sib.popularlibraries.mvp.model.repo.IGitUserRepos
 import io.reactivex.rxjava3.core.Single
@@ -14,29 +13,18 @@ import io.reactivex.rxjava3.schedulers.Schedulers
 class RetrofitTheUserRepos(
     private val api: IDataSource,
     private val networkStatus: INetworkStatus,
-    private val db: Database
+    private val cache: IGithubReposCache
+    //private val db: Database
 ): IGitUserRepos {
     override fun getUserRepos(user: GithubUser): Single<List<GithubRepository>> =
         networkStatus.isOnlineSingle().flatMap { isOnline ->
             if (isOnline) {
                 api.getUserReposByLogin(user.login)
-                    .flatMap { repositories ->
-                        Single.fromCallable {
-                            val roomUser = db.userDao.findByLogin(user.login)
-                            val roomRepos = repositories.map {
-                                RoomGithubRepository(it.id, it.name, it.forksCount, roomUser?.id ?: 0L) }
-                            db.repositoryDao.insert(roomRepos)
-                            repositories
-                        }
+                    .flatMap {
+                        cache.setRepos(it, user)
                     }
             } else {
-                Single.fromCallable {
-                    val roomUser = db.userDao.findByLogin(user.login)
-                    roomUser?.run{
-                        db.repositoryDao.findForUser(id).map { GithubRepository(it.id, it.name, it.forksCount) }
-                    } ?: throw RuntimeException("No such user in cache")
-                }
-
+                cache.getRepos(user)
             }
         }.subscribeOn(Schedulers.io())
 }
