+++ app/src/main/java/com/nick_sib/popularlibraries/mvp/view/image/GlideImageLoader.kt
@@ -2,14 +2,13 @@ package com.nick_sib.popularlibraries.mvp.view.image
 
 
 import android.graphics.Bitmap
-import com.bumptech.glide.request.target.Target
 import android.widget.ImageView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.DataSource
 import com.bumptech.glide.load.engine.GlideException
 import com.bumptech.glide.request.RequestListener
+import com.bumptech.glide.request.target.Target
 import com.nick_sib.popularlibraries.mvp.model.cache.IUserAvatarCache
-import com.nick_sib.popularlibraries.mvp.model.entity.GithubUser
 
 class GlideImageLoader(private val cache: IUserAvatarCache) : IImageLoader<ImageView> {
     override fun loadInto(url: String, container: ImageView, userLogin: String) {
